+++ changelog.md
@@ -1,4 +1,4 @@
-[![Downloads](https://img.shields.io/github/downloads/zevnda/steam-game-idler/1.7.0-beta.8/total?style=for-the-badge&logo=github&color=137eb5)](https://github.com/zevnda/steam-game-idler/releases/download/1.7.0-beta.8/Steam.Game.Idler_1.7.0.8_x64_en-US.msi)
+[![Downloads](https://img.shields.io/github/downloads/zevnda/steam-game-idler/1.7.0-beta.9/total?style=for-the-badge&logo=github&color=137eb5)](https://github.com/zevnda/steam-game-idler/releases/download/1.7.0-beta.9/Steam.Game.Idler_1.7.0.9_x64_en-US.msi)
 
 > [!Important]
 > ### This is a beta release
@@ -7,8 +7,8 @@
 > This release will not receive automatic updates. If a newer beta version becomes available you will need to install it manually.
 
 ## Changelog
-- Added a chat feature which can be accessed via the sidebar
-  - Jury is still out on if this makes it to the stable release of `1.7.0`. I guess you can let me know in the chat if it should stay or go 😅
+- All beta users who use the chat feature between now and the release of stable version `1.7.0` will get an `early supporter` role with an exclusive badge next to their username
+- Improved the styling of the chat
 
 ## Issues
 If you have any issues while using this beta version, please [open a new issue](https://github.com/zevnda/steam-game-idler/issues/new?template=issue_report.yml)